Add files via upload
I built this dashboard using Microsoft Excel.
The analysis takes a deep dive into the sales data of Adidas for 2020 and 2021, highlights the positive trajectory in sales and identifies the potential factors contributing to this optimistic forecast.
The year 2021 had greater sales than 2020 by a whooping difference of 535,740,775. The highest-grossing and best-selling product was the Men's Street Footwear. The highest-grossing retailer was West Gear with total sales of 242,964,333.
The sales data for Adidas over the past two years indicates a consistent and healthy performance. The analysis projects a significant upward trend in sales for the coming months. The upward trajectory represents a commendable growth pattern and a positive outlook for the company’s sales.
Based on this forecast, inventory planning, enhanced marketing strategies, and discounted prices on low-grossing products could further boost sales in the coming months.